services/generator/requirements.txt
psycopg2-binary==2.9.9
Faker==24.0.0
